renamed everything in routes folder from 'forum'->'thread'
