+++ routes/index.js
@@ -1,14 +1,14 @@
 //setup router
 const router = require('express').Router();
 //import js files performing the work on the route
-const getForums = require('./getForums');
-const createForum = require('./createForum');
-const updateForum = require('./updateForum');
-const deleteForum = require('./deleteForum');
+const getThreads = require('./getThreads');
+const createThread = require('./createThread');
+const updateThread = require('./updateThread');
+const deleteThread = require('./deleteThread');
 //connect the js files to the route
-router.get("/forum",getForums);
-router.post("/forum",createForum);
-router.put("forum",updateForum);
-router.delete("forum",deleteForum);
+router.get("/thread",getThreads);
+router.post("/thread",createThread);
+router.put("/thread/:id",updateThread);
+router.delete("/thread/:id",deleteThread);
 //export router to 
 module.exports = router;
